@@ -1,6 +1,6 @@
 module tophatdemon.com/total-invasion-ii
 
-go 1.23.0
+go 1.24.1
 
 require (
 	github.com/go-gl/gl v0.0.0-20211210172815-726fda9656d6
